Bug fixes, random intervals, and cleanup
Made it so the script now waits for a random amount of frames before pressing a button, and also holds the button for a random amount of frames before releasing it. Removed an unnecessary callback that caused the random inputs to not be correctly handled. Added a seed to the RNG so that the script does not play the same set of inputs every time it's loaded. Added comments and cleaned up variables a bit.